Preserve string charset (Ruby encoding) round trip.
When the incoming body string has an unknown or incomplete (dummy?) character
set, then some workarounds are implemented to get Regexp's working,
and round trip encoding is not supported.

The string conversion is done in Body, but the information about the encoding
of the email is only in Message/Part, therefore attach that information to the
Body *only* when the body is a text that has a valid string encoding.